refactored topics
