+++ src/content/index.ts
@@ -1,0 +1,6 @@
+import { getting_started } from './getting_started';
+import { javascript_refresher } from './javascript_refresher';
+export const Topics: any = {
+	getting_started,
+	javascript_refresher,
+};
